Fix issue with "inmanta module build" command on a v1 module if inmanta_plugins dir already exists (Issue #4954, PR #4994)
Pull request opened by the merge tool on behalf of #4994

+++ src/inmanta/moduletool.py
@@ -1291,6 +1291,12 @@ class ModuleConverter:
         # move plugins or create
         old_plugins = os.path.join(output_directory, "plugins")
         new_plugins = os.path.join(output_directory, "inmanta_plugins", self._module.name)
+        if os.path.exists(new_plugins) and os.listdir(new_plugins):
+            raise ModuleBuildFailedError(
+                msg=f"Could not build module: inmanta_plugins/{self._module.name} directory already exists and is not empty"
+            )
+        if os.path.exists(new_plugins):
+            os.rmdir(new_plugins)
         if os.path.exists(old_plugins):
             shutil.move(old_plugins, new_plugins)
         else:
